+++ objectified-ui/public/WHATS_NEW.md
@@ -17,6 +17,8 @@ We continue to improve the platform based on your feedback with improvements and
   - Added memory profiling and performance improvements for large canvases
   - Added icons to class nodes for better visual distinction of nodes
   - Added focus mode to allow nodes to be selected, and to show only the associated nodes and relationships
+    - Focus mode can be toggled at any time, toggled off by hitting [ESC]
+    - Focus mode now blurs non-focused nodes
   - Increased zoom out resolution to 10%
   - Nodes now show property references even when not expanded
   - Improved search to include properties and descriptions in the results